+++ astroquery/irsa/tests/test_irsa.py
@@ -10,10 +10,10 @@ from astropy.table import Table
 import astropy.coordinates as coord
 import astropy.units as u
 
-from ...utils.testing_tools import MockResponse
-from ...utils import commons
-from ... import irsa
-from ...irsa import conf
+from astroquery.utils.testing_tools import MockResponse
+from astroquery.utils import commons
+from astroquery.irsa import Irsa, conf
+from astroquery import irsa
 
 DATA_FILES = {'Cone': 'Cone.xml',
               'Box': 'Box.xml',
@@ -35,11 +35,11 @@ def patch_get(request):
         mp = request.getfixturevalue("monkeypatch")
     except AttributeError:  # pytest < 3
         mp = request.getfuncargvalue("monkeypatch")
-    mp.setattr(irsa.Irsa, '_request', get_mockreturn)
+    mp.setattr(Irsa, '_request', get_mockreturn)
     return mp
 
 
-def get_mockreturn(method, url, params=None, timeout=10, **kwargs):
+def get_mockreturn(method, url, params=None, timeout=10, cache=False, **kwargs):
     filename = data_path(DATA_FILES[params['spatial']])
     content = open(filename, 'rb').read()
     return MockResponse(content, **kwargs)
@@ -78,26 +78,26 @@ def test_parse_coordinates(coordinates, expected):
 
 
 def test_args_to_payload():
-    out = irsa.core.Irsa._args_to_payload("fp_psc")
+    out = Irsa._args_to_payload("fp_psc")
     assert out == dict(catalog='fp_psc', outfmt=3, outrows=conf.row_limit,
                        selcols='')
 
 
 @pytest.mark.parametrize(("coordinates"), OBJ_LIST)
 def test_query_region_cone_async(coordinates, patch_get):
-    response = irsa.core.Irsa.query_region_async(
+    response = Irsa.query_region_async(
         coordinates, catalog='fp_psc', spatial='Cone',
         radius=2 * u.arcmin, get_query_payload=True)
     assert response['radius'] == 2
     assert response['radunits'] == 'arcmin'
-    response = irsa.core.Irsa.query_region_async(
+    response = Irsa.query_region_async(
         coordinates, catalog='fp_psc', spatial='Cone', radius=2 * u.arcmin)
     assert response is not None
 
 
 @pytest.mark.parametrize(("coordinates"), OBJ_LIST)
 def test_query_region_cone(coordinates, patch_get):
-    result = irsa.core.Irsa.query_region(
+    result = Irsa.query_region(
         coordinates, catalog='fp_psc', spatial='Cone', radius=2 * u.arcmin)
 
     assert isinstance(result, Table)
@@ -105,18 +105,18 @@ def test_query_region_cone(coordinates, patch_get):
 
 @pytest.mark.parametrize(("coordinates"), OBJ_LIST)
 def test_query_region_box_async(coordinates, patch_get):
-    response = irsa.core.Irsa.query_region_async(
+    response = Irsa.query_region_async(
         coordinates, catalog='fp_psc', spatial='Box',
         width=2 * u.arcmin, get_query_payload=True)
     assert response['size'] == 120
-    response = irsa.core.Irsa.query_region_async(
+    response = Irsa.query_region_async(
         coordinates, catalog='fp_psc', spatial='Box', width=2 * u.arcmin)
     assert response is not None
 
 
 @pytest.mark.parametrize(("coordinates"), OBJ_LIST)
 def test_query_region_box(coordinates, patch_get):
-    result = irsa.core.Irsa.query_region(
+    result = Irsa.query_region(
         coordinates, catalog='fp_psc', spatial='Box', width=2 * u.arcmin)
 
     assert isinstance(result, Table)
@@ -129,12 +129,9 @@ poly2 = [(10.1 * u.deg, 10.1 * u.deg), (10.0 * u.deg, 10.1 * u.deg),
          (10.0 * u.deg, 10.0 * u.deg)]
 
 
-@pytest.mark.parametrize(("polygon"),
-                         [poly1,
-                          poly2
-                          ])
+@pytest.mark.parametrize(("polygon"), [poly1, poly2])
 def test_query_region_async_polygon(polygon, patch_get):
-    response = irsa.core.Irsa.query_region_async(
+    response = Irsa.query_region_async(
         "m31", catalog="fp_psc", spatial="Polygon",
         polygon=polygon, get_query_payload=True)
 
@@ -145,7 +142,7 @@ def test_query_region_async_polygon(polygon, patch_get):
             b1 = float(b1)
             np.testing.assert_almost_equal(a1, b1)
 
-    response = irsa.core.Irsa.query_region_async(
+    response = Irsa.query_region_async(
         "m31", catalog="fp_psc", spatial="Polygon", polygon=polygon)
 
     assert response is not None
@@ -156,7 +153,7 @@ def test_query_region_async_polygon(polygon, patch_get):
                           poly2,
                           ])
 def test_query_region_polygon(polygon, patch_get):
-    result = irsa.core.Irsa.query_region(
+    result = Irsa.query_region(
         "m31", catalog="fp_psc", spatial="Polygon", polygon=polygon)
 
     assert isinstance(result, Table)
@@ -166,7 +163,7 @@ def test_query_region_polygon(polygon, patch_get):
                          zip(('Cone', 'Box', 'Polygon', 'All-Sky'),
                              ('Cone', 'Box', 'Polygon', 'NONE')))
 def test_spatial_valdi(spatial, result):
-    out = irsa.core.Irsa._parse_spatial(
+    out = Irsa._parse_spatial(
         spatial, coordinates='m31', radius=5 * u.deg, width=5 * u.deg,
         polygon=[(5 * u.hour, 5 * u.deg)] * 3)
     assert out['spatial'] == result
@@ -176,4 +173,4 @@ def test_spatial_valdi(spatial, result):
                                         'All-sky', 'invalid', 'blah')])
 def test_spatial_invalid(spatial):
     with pytest.raises(ValueError):
-        irsa.core.Irsa._parse_spatial(spatial, coordinates='m31')
+        Irsa._parse_spatial(spatial, coordinates='m31')